+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.0.0 (2018-08-10)
+
 * Support Handoff from app to Safari
 * Rename `configuration` to `webViewConfiguration`
 * Update SwiftLint 0.26.0